requirements.txt
beautifulsoup4==4.8.1
bleach==3.1.0
Flask-SQLAlchemy==2.4.1
Flask==1.1.1
gunicorn==19.9.0
lxml==4.4.1
psycopg2-binary==2.8.4
python-twitter==3.5
requests==2.22.0
selenium==3.141.0
slackclient==2.9.3
